Unfuck definitions of primitives and such, slightly, along with a few miscellaneous changes to the interpreter.
Spice up define-kernel-primitive to make definitions easier.
Particularly, the env parameter is now declared ignorable, it has
parseable (of depth one) arglists, and the artlists can have types in
them.  This eliminates the need for a bunch of (declare (ignore env))
and (check-args ...) and (check-type ...) in the human-written code.

Other changes include condition reporting and adding some functions I'll
probably forget to use.